+++ src/el/widgets/ctkex/_button_ex.py
@@ -9,12 +9,15 @@ This source code is licensed under the Apache-2.0 license found in the
 LICENSE file in the root directory of this source tree. 
 """
 
+import math
 import typing
-from .._deps import *
+from copy import copy
 
 from el.observable import MaybeObservable, maybe_observe, maybe_obs_value
 from el.ctk_utils.types import *
 from el.ctk_utils import apply_to_config
+
+from .._deps import *
 
 
 class _CTkButtonPassthroughArgs(typing.TypedDict, total=False):
@@ -47,6 +50,7 @@ class _CTkButtonPassthroughArgs(typing.TypedDict, total=False):
 
 class CTkButtonExPassthroughArgs(_CTkButtonPassthroughArgs, total=False):
     touchscreen_mode: MaybeObservable[bool]
+    round_corner_exclude: tuple[bool, bool, bool, bool]
 
 
 class CTkButtonEx(ctk.CTkButton):
@@ -54,6 +58,7 @@ class CTkButtonEx(ctk.CTkButton):
     def __init__(self,
         master: tk.Misc,
         touchscreen_mode: MaybeObservable[bool] = False,
+        round_corner_exclude: tuple[bool, bool, bool, bool] = (False, False, False, False),
         **kwargs: typing.Unpack[_CTkButtonPassthroughArgs]
     ):
         self._touchscreen_mode = maybe_obs_value(touchscreen_mode)
@@ -62,14 +67,105 @@ class CTkButtonEx(ctk.CTkButton):
             apply_to_config(self, "touchscreen_mode"), 
             initial_update=False,
         )
+        self._round_corner_exclude = round_corner_exclude
         super().__init__(master, **kwargs)
 
+    @typing.override
+    def _draw(self, no_color_updates: bool = False):
+        """ Override drawing method to implement round_corner_exclude """
+        # first we draw the normal parts
+        super()._draw(no_color_updates)
+
+        # then on top of that, we draw additional rectangles to cover up some
+        # of the corners that we want to "exclude" from rounding.
+        # First we calculate the width and height the same it is done in ctk_button.py and drawing_engine.py
+        width = self._apply_widget_scaling(self._current_width)
+        if self._round_width_to_even_numbers:
+            width = math.floor(width / 2) * 2  # round (floor) _current_width and _current_height and restrict them to even values only
+        height = self._apply_widget_scaling(self._current_height)
+        if self._round_height_to_even_numbers:
+            height = math.floor(height / 2) * 2
+        border_width = round(self._apply_widget_scaling(self._border_width))
+
+        requires_recoloring = False
+
+        for i, exclude in enumerate(self._round_corner_exclude):
+            # create border and inner rectangles to cover the selected corners
+            if exclude:
+                if not self._canvas.find_withtag(f"corner_cover_border_{i}"):
+                    self._canvas.create_rectangle(
+                        0, 0, 0, 0,
+                        tags=(
+                            f"corner_cover_border_{i}",
+                            "corner_cover_border_parts",
+                            "border_parts",
+                        ),
+                        width=0,
+                    )
+                    requires_recoloring = True
+                if not self._canvas.find_withtag(f"corner_cover_inner_{i}"):
+                    self._canvas.create_rectangle(
+                        0, 0, 0, 0,
+                        tags=(
+                            f"corner_cover_inner_{i}",
+                            "corner_cover_inner_parts",
+                            "inner_parts",
+                        ),
+                        width=0,
+                    )
+                    requires_recoloring = True
+
+                # position them properly
+                self._canvas.coords(f"corner_cover_border_{i}", (
+                    width if 1 <= i <= 2 else 0, 
+                    height if i >= 2 else 0,
+                    width / 2, height / 2
+                ))
+                self._canvas.coords(f"corner_cover_inner_{i}", (
+                    (width - border_width) if 1 <= i <= 2 else border_width,
+                    (height - border_width) if i >= 2 else border_width,
+                    width / 2, height / 2
+                ))
+
+            # or if this corner is not needed (anymore), make sure it's parts are definitely deleted
+            else:
+                self._canvas.delete(f"corner_cover_border_{i}", f"corner_cover_inner_{i}")
+        
+        # raise the elements to ensure the cover parts are always on the top and
+        # the inner parts are always above the border parts. Without this, visual
+        # artifacts may occur after a redraw when the elements already exist.
+        self._canvas.tag_raise("corner_cover_border_parts")
+        self._canvas.tag_raise("corner_cover_inner_parts")
+
+        # when parts are created, we also need to re-run the color setting procedure
+        # done by super()._draw() to set the colors of the corner covers as well.
+        # This has been copied from ctk_button.py.
+        if no_color_updates is False or requires_recoloring:
+            # set color for the button border parts (outline)
+            self._canvas.itemconfig("border_parts",
+                                    outline=self._apply_appearance_mode(self._border_color),
+                                    fill=self._apply_appearance_mode(self._border_color))
+            
+            # set color for inner button parts
+            if self._fg_color == "transparent":
+                self._canvas.itemconfig("inner_parts",
+                                        outline=self._apply_appearance_mode(self._bg_color),
+                                        fill=self._apply_appearance_mode(self._bg_color))
+            else:
+                self._canvas.itemconfig("inner_parts",
+                                        outline=self._apply_appearance_mode(self._fg_color),
+                                        fill=self._apply_appearance_mode(self._fg_color))
+
+            #self._canvas.itemconfig("corner_cover_border_parts",
+            #                        fill="orange" if self._desired_width == self._current_width else "red")
+            #self._canvas.itemconfig("corner_cover_inner_parts",
+            #                        fill="lime" if self._desired_width == self._current_width else "green")
 
     @typing.override
     def _clicked(self, event=None):
         if self._state != tk.DISABLED:
             # click animation: change color with .on_leave() and back to normal after 100ms with click_animation()
-            # edit in ctkex: If hovering is disabled or we are in touchscreen mode, we invert the animation 
+            # edit in ctkex: If hovering is disabled or we are in touchscreen mode, we invert the animation
             # so click is still visible (_click_animation is also modified)
             self._click_animation_running = True    # important: must be set before _on_enter so it renders
             if self._touchscreen_mode or not self._hover:
@@ -94,7 +190,7 @@ class CTkButtonEx(ctk.CTkButton):
             # we reset to previous state as hover may have been enabled
             # and we were just in touchscreen mode, so we don't want to force it off
             self._hover = hover_before
-        
+
         # if we are not in animation AND not in touchscreen mode we run the regular
         # _on_enter handler. That one will then check for whether hover is enabled or not.
         elif not self._touchscreen_mode:
@@ -128,7 +224,7 @@ class CTkButtonEx(ctk.CTkButton):
             self._touchscreen_mode = maybe_obs_value(kwargs["touchscreen_mode"])
             kwargs.pop("touchscreen_mode")
             self._set_cursor()
-        
+
         super().configure(require_redraw, **kwargs)
 
     @typing.override
@@ -137,7 +233,7 @@ class CTkButtonEx(ctk.CTkButton):
             return self._touchscreen_mode
         else:
             return super().cget(attribute_name)
-    
+
     @typing.override
     def _set_cursor(self):
         """ Override this to allow for disable cursor in touchscreen mode """